Defer connecting to database until actually needed
Previously, uploader would try to connect to the local and remote
databases. If it was unable to connect, then it would retry a certain
amount of time. As a result, it would block home assistant from booting
up! I realized that the whole retrying thing is stupid. Instead, the
uploader component starts no matter what. When it tries to download and
upload data, it tries to connect to the database. If it can't it retries
 again based on the back up interval. This way, the uploader will run
all the time, no matter what!