Fix device clean up
I was deleting devices that haven't been seen for awhile. This was
causing problems in Home Assistant. When the device was rediscovered
it would create a new entity ID for it (appending a 2 to the end of it).
This would create a new entry in the database. Now instead of deleting
the device, I just stop trying to get data from it until it is
discovered again.